Fix #376813 – problem with filenames that contain whitespace ...  	Partly fix #374700: still missing tango icons  	Fix libburn capabilities about multisession  	Marked some strings for translations  	Removed unneeded dialogs

+++ src/burn-iso9660.c
@@ -133,14 +133,14 @@ brasero_iso9660_is_primary_descriptor (const char *buffer,
 
 gboolean
 brasero_iso9660_get_size (const gchar *block,
-			  gint32 *nb_blocks,
+			  gint64 *nb_blocks,
 			  GError **error)
 {
 	BraseroIsoPrimary *vol;
 
 	/* read the size of the volume */
 	vol = (BraseroIsoPrimary *) block;
-	*nb_blocks = brasero_iso9660_get_733_val (vol->vol_size);
+	*nb_blocks = (gint64) brasero_iso9660_get_733_val (vol->vol_size);
 
 	return TRUE;
 }
