+++ src/burn-volume.h
@@ -67,7 +67,7 @@ brasero_volume_is_iso9660 (const gchar *path,
 
 gboolean
 brasero_volume_get_size (const gchar *path,
-			 gint32 *nb_blocks,
+			 gint64 *nb_blocks,
 			 GError **error);
 gboolean
 brasero_volume_get_label (const gchar *path,
@@ -76,7 +76,7 @@ brasero_volume_get_label (const gchar *path,
 BraseroVolFile *
 brasero_volume_get_files (const gchar *path,
 			  gchar **label,
-			  gint32 *nb_blocks,
+			  gint64 *nb_blocks,
 			  GError **error);
